Added a very basic chaining middleware implementation
